makes headers consistent between menu and home and exports header to component for use in other pages
