+++ tailwind.config.js
@@ -36,6 +36,7 @@ module.exports = {
         '17': '4.25rem', // 68px
         '18': '4.5rem', // 72px
         '20': '5rem', // 80px
+        '25': '6.25rem', // 100px
         '27': '6.75rem', // 108px
         '30': '7.5rem', // 120px
         '36': '9rem', // 144px
