+++ views/Home.js
@@ -6,20 +6,17 @@ import { useRouteDispatch } from '../state/routeContext'
 import { useFoodDataDispatch } from '../state/foodDataContext'
 import {
   CHANGE_VIEW,
-  MENU,
   // LOADING,
   FOOD_DATA,
   ERROR,
   SET_FOOD_PHOTO,
 } from '../utils/constants'
+import { Header } from '../components/Header'
 import fileNameFormatter from '../utils/fileNameFormatter'
 
 import buttonBG from '../public/backgrounds/camera_bg.svg'
-import menuBG from '../public/backgrounds/menu_bg.svg'
 import cameraBG from '../public/icons/camera_icn.svg'
 
-import menu from '../public/icons/menu.svg'
-import logo from '../public/logos/logo2.svg'
 import cameraButton from '../public/icons/btn_round.svg'
 
 const Home = () => {
@@ -60,15 +57,7 @@ const Home = () => {
 
   return (
     <Dashboard>
-      <MenuContainer>
-        <Logo alt="logo" src={logo} />
-        <div
-          onClick={() => routeDispatch({ type: CHANGE_VIEW, view: MENU })}
-          onKeyPress={() => routeDispatch({ type: CHANGE_VIEW, view: MENU })}
-        >
-          <img alt="menu" src={menu} />
-        </div>
-      </MenuContainer>
+      <Header />
       <CameraContainer>
         <p className="font-bold">What is your child eating for dinner?</p>
         <p>Share a photo</p>
@@ -121,19 +110,5 @@ const ButtonForm = styled.form.attrs({
 const CameraContainer = styled.div.attrs({
   className: 'w-full font-xl px-4 pt-5 pb-2d5',
 })``
-
-const MenuContainer = styled.div.attrs({
-  className: 'w-full px-6 pt-5d5',
-})`
-  display: grid;
-  grid-template-columns: 1fr 12.5%;
-  background: url(${menuBG}) left top/cover no-repeat;
-`
-
-const Logo = styled.img.attrs({
-  className: 'mr-12 w-20 sm:w-auto',
-})`
-  justify-self: end;
-`
 
 export default Home
